Stego/requirements.txt

gradio
Pillow
